refactoring - put modals to views/components for further use from other views
