+++ src/store.js
@@ -1,6 +1,6 @@
 import {createStore, applyMiddleware, compose} from 'redux'
 import thunk from 'redux-thunk'
-import rootReducer from './services/reducers'
+import rootReducer from './services/reducer'
 
 let initialState = {};
