+++ src/views/PublicRouter/services/reducer.js
@@ -1,9 +1,7 @@
 import { combineReducers } from 'redux'
-import publicReducer from './publicReducer'
-import homeReducer from '../../views/Home/services/reducer'
+import homeReducer from '../views/Home/services/reducer'
 
 // here we collect ALL reducers from views directory
 export default combineReducers({
-	publicReducer,
 	homeReducer
 })
